lab2/file-service/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pyjwt